Add files via upload
The app.js file is javascript file, I have made this file using vue.js and also javascript code is in comment so you can see the difference in code for javascript and vuejs. Other are the files for html and css. I had to make some changes in code in html file for vue but not in the css file